requirements.txt
flappy_bird_gymnasium==0.4.0
gymnasium==0.29.1
matplotlib==3.7.2
numpy==2.0.0
PyYAML==6.0.1
torch==2.0.1
